feat(#229): fix all qulice suggestions
