+++ src/main/java/org/eolang/opeo/compilation/XmirParser.java
@@ -36,8 +36,8 @@ import org.eolang.opeo.ast.ArrayConstructor;
 import org.eolang.opeo.ast.AstNode;
 import org.eolang.opeo.ast.Attributes;
 import org.eolang.opeo.ast.Cast;
-import org.eolang.opeo.ast.ClassConstant;
 import org.eolang.opeo.ast.ClassField;
+import org.eolang.opeo.ast.ClassName;
 import org.eolang.opeo.ast.Constructor;
 import org.eolang.opeo.ast.ConstructorDescriptor;
 import org.eolang.opeo.ast.Field;
@@ -183,7 +183,7 @@ final class XmirParser {
         } else if ("long".equals(base)) {
             result = new Literal(node);
         } else if ("type".equals(base)) {
-            result = new ClassConstant(node);
+            result = new ClassName(node);
         } else if (".super".equals(base)) {
             final List<XmlNode> inner = node.children().collect(Collectors.toList());
             final AstNode instance = this.node(inner.get(0));
